Fixed wrong type for threadID (Changed to uint16_t from original uint8_t)

+++ src/RevExt.cc
@@ -35,7 +35,7 @@ void RevExt::SetOTable(std::vector<RevInstEntry> InstVect){
   otable = InstVect;
 }
 
-bool RevExt::Execute(unsigned Inst, RevInst payload, uint8_t threadID){
+bool RevExt::Execute(unsigned Inst, RevInst payload, uint16_t threadID){
 
   // ensure that the target instruction is within scope
   if( Inst > (table.size()-1) ){
